General: Add some niceties + Server: Inherit from ThreadingTCPServer
This commit:
  * Adds a pip dependency on black, for formatting + linting
  * Re-formats all code according to black default settings
  * Adds a pip dependency on mypy, for type checking
  * Fixes a couple of type errors
  * Adds __init__.py
  * Changes GeminiServer to inherit from ThreadingTCPServer rather than
    using composition. I'll probably regret this later. But this does
    remove the previous ugly practice of adding attributes to a
    standard library class.

Note that there is no type stub file for the colorama package. Until
either I or someone else adds this, mypy has a single error for that. To
fix this, run as:
  mypy *.py --ignore-missing-imports
This could silence other useful errors, so an unfortunate best practice
is to run without `--ignore-missing-improts` to confirm missing stub
files are the only issues; then re-run with it.